ready to tag v10r3p7: fix many unix warnings

+++ TkrRecon/src/PatRec/Combo/ComboFindTrackTool.cxx
@@ -871,7 +871,7 @@ bool ComboFindTrackTool::incorporate(Candidate* trial)
     } 
 
     // we're full up... one of these candidates has got to go...
-    if(m_candidates.size()==m_maxCandidates) {
+    if(m_candidates.size()==(unsigned int)m_maxCandidates) {
         iterator last = --end();
         Candidate* lastCand = *last;
         // just skip it if it's no better than the last one
